final bugfixes

+++ src/playground/discochal/removeif0.sh
@@ -1,6 +1,7 @@
-#/bin/bash
-FILESIZE=$(du -k "$1" | awk '{print $1}')
-
-if [ "$FILESIZE" -eq "0" ]; then
-  rm $1
+if [ -f "$1" ] 
+then
+	FILESIZE=$(du -k "$1" | awk '{print $1}')
+	if [ "$FILESIZE" -eq "0" ]; then
+  	   rm $1
+	fi
 fi
