+++ src/playground/discochal/removeuser.sh
@@ -1,5 +1,8 @@
-cat watched.cfg | grep -v $1 > watched.new
-mv watched.new watched.cfg
+if [ -f "watched.cfg" ]
+then 
+	cat watched.cfg | grep -v $1 > watched.new
+	mv watched.new watched.cfg
+fi
 
 if [ "$2" != "nooutput"  ] 
 then
